[Fix] fix auth module api issues and changes and and fix design issues.
