+++ lib/data/network/network_api_service.dart
@@ -1,33 +1,64 @@
+import 'dart:developer';
 import 'dart:io';
 
 import 'package:dio/dio.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 import 'package:table_menu_customer/data/network/base_api_service.dart';
 import 'package:table_menu_customer/res/services/api_endpoints.dart';
+import 'package:table_menu_customer/utils/helpers.dart';
 
 import '../app_exceptions.dart';
-import '../interceptors/auth_interceptor.dart';
-import '../interceptors/error_interceptor.dart';
 
 class NetworkApiService extends BaseApiService {
-  final Dio _dio = Dio(BaseOptions(
+
+  static final Dio _dio = Dio(BaseOptions(
       baseUrl: ApiEndPoint.baseUrl,
+      sendTimeout: const Duration(milliseconds: 30000),
+      receiveTimeout: const Duration(milliseconds: 60000),
+      connectTimeout: const Duration(milliseconds: 30000),
       followRedirects: false,
       validateStatus: (status) {
+        log("@@status : $status");
         return status! < 500;
       }));
 
-  NetworkApiService() {
-    _dio.interceptors.add(AuthInterceptor());
-    _dio.interceptors.add(ErrorInterceptor());
+
+  setupInterceptors() {
+    _dio.interceptors.add(
+      InterceptorsWrapper(
+        onRequest: (options, handler) {
+          return handler.next(options);
+        },
+        onResponse: (response, handler) {
+          if (response.statusCode == 401) {
+            RequestOptions originalRequest = response.requestOptions;
+            // Throw an error to trigger the onError callback
+            throw DioException(
+              requestOptions: originalRequest,
+              response: response,
+              error: 'Unauthorized',
+              type: DioExceptionType.badResponse,
+            );
+          }
+          // Add your custom logic here for response interception
+          return handler.next(response);
+        },
+        onError: (DioException error, handler) async {
+          if (error.response?.statusCode == 401) {
+            // TODO perform force logout
+          }
+          // Handle other errors if needed
+          return handler.next(error);
+        },
+      ),
+    );
   }
 
   // to make HTTP requests to an API.
   @override
   Future<Response> getGetApiResponse(String url) async {
     try {
-      final prefs = await SharedPreferences.getInstance();
-      final token = prefs.getString('token');
+      final token = await getToken();
 
       var headers = {
         'Authorization': 'Token $token',
@@ -54,15 +85,14 @@ class NetworkApiService extends BaseApiService {
   Future<Response> getGetApiResponseWithParams(
       String url, String params) async {
     try {
-      final prefs = await SharedPreferences.getInstance();
-      final token = prefs.getString('token');
-
+      final token = await getToken();
       var headers = {
         'Authorization': 'Token $token',
         'Accept': 'application/json'
       };
       var queryParams = {'query': params};
-
+      log("queryParams $queryParams");
+      log("headers $headers");
       final response = await _dio
           .get(url,
               options: Options(
@@ -70,10 +100,7 @@ class NetworkApiService extends BaseApiService {
               ),
               queryParameters: queryParams)
           .timeout(const Duration(seconds: 10));
-      //print(response.data);
       return response;
-      // responseJson = returnResponse(response);
-      // return responseJson;
     } on SocketException {
       throw FetchDataExceptions(
           'Error Occured While Communicating with Server');
@@ -85,12 +112,10 @@ class NetworkApiService extends BaseApiService {
   Future<Response> getAuthApiResponse(String url, dynamic data,
       {String? verifyToken}) async {
     try {
-      print("token $verifyToken");
       var headers = {
         'HTTP_AUTHORIZATION': 'Token $verifyToken',
         'Accept': 'application/json'
       };
-      print(headers);
       var response = await _dio
           .post(url,
               data: data,
@@ -100,7 +125,6 @@ class NetworkApiService extends BaseApiService {
                   contentType: Headers.jsonContentType,
                   responseType: ResponseType.json))
           .timeout(const Duration(seconds: 10));
-      print(response.statusCode);
       return response;
     } on SocketException {
       throw FetchDataExceptions(
@@ -112,8 +136,7 @@ class NetworkApiService extends BaseApiService {
   @override
   Future<Response> getPostApiResponse(String url, dynamic data) async {
     try {
-      final prefs = await SharedPreferences.getInstance();
-      final token = prefs.getString('token');
+      final token = await getToken();
       var headers = {
         'Authorization': 'Token $token',
         'Accept': 'application/json'
@@ -140,8 +163,7 @@ class NetworkApiService extends BaseApiService {
   @override
   Future<Response> getPatchApiResponse(String url, dynamic data) async {
     try {
-      final prefs = await SharedPreferences.getInstance();
-      final token = prefs.getString('token');
+      final token = await getToken();
 
       var headers = {
         'Authorization': 'Token $token',
@@ -170,9 +192,7 @@ class NetworkApiService extends BaseApiService {
   @override
   Future<Response> getPutApiResponse(String url, dynamic data) async {
     try {
-      final prefs = await SharedPreferences.getInstance();
-      final token = prefs.getString('token');
-
+      final token = await getToken();
       var headers = {
         'Authorization': 'Token $token',
         'Accept': 'application/json'
@@ -200,8 +220,7 @@ class NetworkApiService extends BaseApiService {
   @override
   Future<Response> getDeleteApiResponse(String url) async {
     try {
-      final prefs = await SharedPreferences.getInstance();
-      final token = prefs.getString('token');
+      final token = await getToken();
 
       var headers = {
         'Authorization': 'Token $token',
