+++ lib/utils/constants/constants_text.dart
@@ -1,3 +1,4 @@
-class Constants {
-
-}
+const double BORDER_RADIUS = 15.0;
+const fontSemiBold = 'Poppins-SemiBold';
+const fontMedium = 'Poppins-Medium';
+const fontRegular = 'Poppins-Regular';
