+++ lib/utils/font/text_style.dart
@@ -1,0 +1,52 @@
+// ignore_for_file: file_names
+
+import 'package:flutter/material.dart';
+import 'package:table_menu_customer/utils/constants/constants_text.dart';
+
+TextStyle titleTextStyle = const TextStyle(
+  fontFamily: fontMedium, // weight: 500
+  fontSize: 24,
+  fontStyle: FontStyle.normal,
+);
+TextStyle smallTitleTextStyle = const TextStyle(
+  fontFamily: fontMedium, // weight: 500
+  fontSize: 22,
+  fontStyle: FontStyle.normal,
+);
+
+TextStyle textBodyStyle = const TextStyle(
+  fontFamily: fontSemiBold, //  weight: 600
+  fontSize: 16,
+  fontStyle: FontStyle.normal,
+);
+
+TextStyle textRegularStyle = const TextStyle(
+  fontFamily: fontRegular, // weight: 400
+  fontSize: 16,
+  fontStyle: FontStyle.normal,
+);
+TextStyle textSmallRegularStyle = const TextStyle(
+  fontFamily: fontRegular, // weight: 400
+  fontSize: 14,
+  fontStyle: FontStyle.normal,
+);
+TextStyle smallRegularStyle = const TextStyle(
+  fontFamily: fontRegular, // weight: 400
+  fontSize: 12,
+  fontStyle: FontStyle.normal,
+);
+TextStyle textMediumStyle = const TextStyle(
+  fontFamily: fontMedium, // weight: 500
+  fontSize: 16,
+  fontStyle: FontStyle.normal,
+);
+TextStyle textSmallMediumStyle = const TextStyle(
+  fontFamily: fontMedium, // weight: 500
+  fontSize: 14,
+  fontStyle: FontStyle.normal,
+);
+TextStyle textSmallStyle = const TextStyle(
+  fontFamily: fontRegular, // weight: 400
+  fontSize: 14,
+  fontStyle: FontStyle.normal,
+);
