+++ lib/utils/helpers.dart
@@ -1,0 +1,9 @@
+
+
+import 'package:shared_preferences/shared_preferences.dart';
+
+getToken() async {
+  final prefs = await SharedPreferences.getInstance();
+  final token = prefs.getString('token');
+  return token;
+}
