+++ lib/view_model/auth_provider.dart
@@ -5,6 +5,7 @@ import 'package:firebase_messaging/firebase_messaging.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:shared_preferences/shared_preferences.dart';
+import 'package:table_menu_customer/data/network/network_api_service.dart';
 import 'package:table_menu_customer/model/auth_model.dart';
 import 'package:table_menu_customer/model/custom_result_model.dart';
 import 'package:table_menu_customer/model/user_model.dart';
@@ -139,9 +140,12 @@ class AuthProvider extends ChangeNotifier {
     return CustomResultModel(status: false, message: "An error occurred");
   }
 
-  Future<CustomResultModel?> verifyUser(String otp) async {
+  Future<CustomResultModel?> verifyUser(String otp,String email) async {
     setLoading(true);
-    Map<String, dynamic> data = {'otp': otp};
+    Map<String, dynamic> data = {
+      "email": email,
+      "otp": otp
+    };
     var result = await _authRepository.verifyUser(data);
 
     if (result.data['status'] == true) {
@@ -227,8 +231,6 @@ class AuthProvider extends ChangeNotifier {
     var result = await _userInfoRepository.saveUserInfo(formData);
 
     if (result.data['status'] == true) {
-      final prefs = await SharedPreferences.getInstance();
-      await prefs.setBool('userDataBool', true);
       setLoading(false);
       return CustomResultModel(status: true, message: result.data["message"]);
     } else if (result.data['status'] == "False") {
@@ -321,10 +323,10 @@ class AuthProvider extends ChangeNotifier {
     SharedPreferences preferences =
         await SharedPreferences.getInstance();
     await preferences.remove('token');
+    await preferences.remove('verify_pass_token');
   }
 
-
-  void forceLogout(BuildContext context) {
+  void logout(BuildContext context) {
     clearToken();
     // Navigate to the login screen
     Navigator.pushNamedAndRemoveUntil(
