+++ pubspec.yaml
@@ -56,6 +56,7 @@ dependencies:
   flutter_local_notifications: ^15.0.0
   mockito: ^5.4.1
   flutter_slidable: ^3.0.0
+  get_it: ^7.6.0
 
 dev_dependencies:
   flutter_test:
